chore: implemented id verification

+++ src/services/users/controllers/user-controller.js
@@ -1,9 +1,13 @@
 import UserRepositories from '../repositories/user-repositories.js';
 import response from '../../../utils/response.js';
 import InvariantError from '../../../exceptions/invariant-error.js';
+import AuthenticationError from '../../../exceptions/authentication-error.js';
 
 export const viewProfile = async (req, res, next) => {
   const { id } = req.user;
+
+  const validId = await UserRepositories.verifyUserId(id);
+  if (!validId) return next(new AuthenticationError('Invalid access token'));
 
   try {
     const user = await UserRepositories.getUser(id);
@@ -18,6 +22,9 @@ export const editProfile = async (req, res, next) => {
   const { id } = req.user;
   const { name } = req.validated;
 
+  const validId = await UserRepositories.verifyUserId(id);
+  if (!validId) return next(new AuthenticationError('Invalid access token'));
+
   try {
     const newData = await UserRepositories.updateUser(id, name);
     return response(res, 200, 'Profile updated successfully', newData);
@@ -29,6 +36,9 @@ export const editProfile = async (req, res, next) => {
 
 export const removeAccount = async (req, res, next) => {
   const { id } = req.user;
+
+  const validId = await UserRepositories.verifyUserId(id);
+  if (!validId) return next(new AuthenticationError('Invalid access token'));
 
   try {
     await UserRepositories.deleteUser(id);
